move components into page
