+++ app/posts/[slug]/loading.js
@@ -1,0 +1,5 @@
+import PostTitle from './post-title'
+
+export default function Loading() {
+	return (<PostTitle>Loading…</PostTitle>)
+}
