+++ app/posts/[slug]/post-preview.js
@@ -1,7 +1,7 @@
 import Link from 'next/link'
-import Avatar from '../components/avatar'
-import DateComponent from '../components/date'
-import CoverImage from './cover-image'
+import Avatar from '../../../components/avatar'
+import DateComponent from '../../../components/date'
+import CoverImage from '../../../components/cover-image'
 
 export default function PostPreview({
   title,
