+++ app/posts/[slug]/post-title.js
