+++ app/posts/[slug]/rich-text-asset.js
